took flush out of IF and it gets set to 0 in the ID stage
I don't think we need the flush in IF because ins+2 only stalls once in IF and then continues on normally in the next cycle in IF no matter what, so the flush variable doesn't affect it. Feel free to change it back if I'm wrong. Still trying to understand the overall roadmap